Property Type enum tweaks

+++ project_world.cpp
@@ -89,9 +89,9 @@ void DrWorld::initializeWorldSettings(QString new_name)
     getComponent(Object_Components::settings)->setIcon(Component_Icons::Settings);
     addPropertyToComponent(World_Components::settings, World_Properties::name, Property_Type::String, new_name,
                            "World Name", "Name of the current world.");
-    addPropertyToComponent(World_Components::settings, World_Properties::game_direction, Property_Type::Double, 0.0,
+    addPropertyToComponent(World_Components::settings, World_Properties::game_direction, Property_Type::Float, 0.0,
                            "Game Direction", "Default direction, in degrees, to load new scenes, 0 - right, 90 - up, 180 - left, 270 - down, etc.");
-    addPropertyToComponent(World_Components::settings, World_Properties::score_multiplier, Property_Type::Double, 1.0,
+    addPropertyToComponent(World_Components::settings, World_Properties::score_multiplier, Property_Type::Float, 1.0,
                            "Score Multiplier", "Value used as multiplier to adjust speed at which distance scoring is calculated.");
 
     addComponent(World_Components::physics, "Physics", "Starting physics settings for current world.", Component_Colors::Orange_Medium, true);
@@ -100,13 +100,13 @@ void DrWorld::initializeWorldSettings(QString new_name)
                            "Use Physics?", "Whether or not physics is turned on in current world.");
     addPropertyToComponent(World_Components::physics, World_Properties::gravity, Property_Type::PointF, QPointF(0.0, 10.0),
                            "Gravity", "Amount of gravity in x and y directions, can be negative.");
-    addPropertyToComponent(World_Components::physics, World_Properties::time_warp, Property_Type::Double, 60.0,
+    addPropertyToComponent(World_Components::physics, World_Properties::time_warp, Property_Type::Float, 60.0,
                            "Time Warp", "Update calculations per second.");
-    addPropertyToComponent(World_Components::physics, World_Properties::friction, Property_Type::Double, 10.0,
+    addPropertyToComponent(World_Components::physics, World_Properties::friction, Property_Type::Float, 10.0,
                            "Friction", "Global friction setting.");
-    addPropertyToComponent(World_Components::physics, World_Properties::drag, Property_Type::Double, 5.0,
+    addPropertyToComponent(World_Components::physics, World_Properties::drag, Property_Type::Float, 5.0,
                            "Velocity Drag", "Velocity drag.");
-    addPropertyToComponent(World_Components::physics, World_Properties::bounce, Property_Type::Double, 5.0,
+    addPropertyToComponent(World_Components::physics, World_Properties::bounce, Property_Type::Float, 5.0,
                            "Bounce", "Global bounce setting.");
 
 }
